+++ c/platform/arduino_avr/include/RtdAcquire.h
@@ -1,0 +1,7 @@
+#ifndef RTD_ACQUIRE_ARDUINO_LIBRARY_H
+#define RTD_ACQUIRE_ARDUINO_LIBRARY_H
+
+#include "rtd_acquire/arduino_avr.hpp"
+#include "rtd_acquire/max31865.h"
+
+#endif